+++ NEWS.md
@@ -21,6 +21,8 @@ multivariate models thanks to Maxime Dahirel. (#834)
 * Fix post-processing of multivariate multilevel models
 when multiple IDs are used for the same grouping factor
 thanks to @lott999. (#835)
+* Store response category names of ordinal models in the
+output of `posterior_predict` again thanks to Mattew Kay. (#838)
 
 
 # brms 2.11.1